src/main/java/uk/tw/energy/service/MeterReadingService.java
package uk.tw.energy.service;

import org.springframework.stereotype.Service;
import uk.tw.energy.domain.ElectricityReading;

import java.time.Instant;
import java.time.LocalDate;
import java.time.ZoneOffset;
import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import java.util.stream.Collectors;

@Service
public class MeterReadingService {

    private final Map<String, List<ElectricityReading>> meterAssociatedReadings;

    public MeterReadingService(Map<String, List<ElectricityReading>> meterAssociatedReadings) {
        this.meterAssociatedReadings = meterAssociatedReadings;
    }

    public Optional<List<ElectricityReading>> getReadings(String smartMeterId) {
        return Optional.ofNullable(meterAssociatedReadings.get(smartMeterId));
    }

    public Optional<List<ElectricityReading>> getPrevNatualWeekReadings(String smartMeterId) {
        Instant lastWeekFrom = LocalDate.now().minusDays(LocalDate.now().getDayOfWeek().getValue() + 6).atStartOfDay().toInstant(ZoneOffset.UTC);
        Instant lastWeekTo = LocalDate.now().minusDays(LocalDate.now().getDayOfWeek().getValue() - 1).atStartOfDay().toInstant(ZoneOffset.UTC);
        return Optional.of(meterAssociatedReadings.get(smartMeterId).stream()
                .filter(reading -> reading.getTime().isAfter(lastWeekFrom) && reading.getTime().isBefore(lastWeekTo)).collect(Collectors.toList()));
    }

    public void storeReadings(String smartMeterId, List<ElectricityReading> electricityReadings) {
        if (!meterAssociatedReadings.containsKey(smartMeterId)) {
            meterAssociatedReadings.put(smartMeterId, new ArrayList<>());
        }
        meterAssociatedReadings.get(smartMeterId).addAll(electricityReadings);
    }
}
